move scrolling info into lyric editor.
