+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditor.cs
@@ -2,6 +2,7 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
@@ -22,6 +23,9 @@ using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.TimeTags;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
+using osu.Game.Rulesets.Timing;
+using osu.Game.Rulesets.UI.Scrolling;
+using osu.Game.Rulesets.UI.Scrolling.Algorithms;
 using osu.Game.Screens.Edit;
 using osu.Game.Skinning;
 using osuTK.Input;
@@ -51,6 +55,9 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 
         [Cached]
         private readonly BlueprintSelectionState blueprintSelectionState = new();
+
+        [Cached(Type = typeof(IScrollingInfo))]
+        private readonly LocalScrollingInfo scrollingInfo = new();
 
         public Bindable<LyricEditorMode> BindableMode { get; } = new();
 
@@ -465,6 +472,19 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
                 default:
                     throw new ArgumentOutOfRangeException(nameof(mode));
             }
+        }
+
+        private class LocalScrollingInfo : IScrollingInfo
+        {
+            public IBindable<ScrollingDirection> Direction { get; } = new Bindable<ScrollingDirection>(ScrollingDirection.Left);
+
+            public IBindable<double> TimeRange { get; } = new BindableDouble(5000)
+            {
+                MinValue = 1000,
+                MaxValue = 10000
+            };
+
+            public IScrollAlgorithm Algorithm { get; } = new SequentialScrollAlgorithm(new List<MultiplierControlPoint>());
         }
     }
 }
